refactor: ChatInput Refactor ImageTag Ref
