+++ client/src/Component/Molecules/Chat/ChatInput/ChatInput.hook.tsx
@@ -1,4 +1,5 @@
-import { useRef } from "react";
+import { postChat } from "@Util/data";
+import { ChangeEvent, ChangeEventHandler, useRef } from "react";
 import ClientSocket from "Socket";
 
 type chatRoomInfo = {
@@ -13,11 +14,27 @@ export const useChatMessageControl = ({ uid, chatRoomId }: chatRoomInfo) => {
     sendChat({ uid, chatRoomId, message: messageRef.current.value });
     messageRef.current.value = "";
   };
-  return { messageRef, sendMessage };
+  const handleEnterPress = (e: React.KeyboardEvent<HTMLInputElement>) => {
+    if (e.code !== "Enter") return;
+    sendMessage();
+  };
+  return { messageRef, sendMessage, handleEnterPress };
 };
 
 type sendChatPropsType = chatRoomInfo & { message: string };
 const sendChat = ({ uid, chatRoomId, message }: sendChatPropsType) => {
   const { socket } = new ClientSocket(uid);
   socket?.emit("sendChat", { chatRoomId, message: { from: uid, message, read: false, source: "" } });
+};
+
+export const useChatImageControl = ({ uid, chatRoomId }: chatRoomInfo) => {
+  const imageInputTag = useRef<HTMLInputElement>(null);
+  const handleImageButtonClick = () => (imageInputTag.current as HTMLInputElement).click();
+
+  const changeImage: ChangeEventHandler<HTMLInputElement> = (event: ChangeEvent<HTMLInputElement>) => {
+    if (!event.target.files) return;
+    postChat(chatRoomId, uid, event.target.files[0]);
+  };
+
+  return { imageInputTag, handleImageButtonClick, changeImage };
 };
